+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Components/EventBoxTooltip.cs
@@ -37,8 +37,11 @@ namespace MonoDevelop.Components
 	public class EventBoxTooltip : IDisposable
 	{
 		EventBox eventBox;
-		string tip;
 		TooltipPopoverWindow tooltipWindow;
+		ImageView image;
+		Pixbuf normalPixbuf;
+		Pixbuf activePixbuf;
+		string tip;
 		bool mouseOver;
 
 		public Atk.Object Accessible {
@@ -51,11 +54,6 @@ namespace MonoDevelop.Components
 		/// The EventBox should have Visible set to false otherwise the tooltip pop window
 		/// will have the wrong location.
 		/// </summary>
-
-		ImageView image;
-		Pixbuf normalPixbuf;
-		Pixbuf activePixbuf;
-
 		public EventBoxTooltip (EventBox eventBox)
 		{
 			this.eventBox = eventBox;
@@ -172,38 +170,5 @@ namespace MonoDevelop.Components
 		public TaskSeverity? Severity { get; set; }
 		public PopupPosition Position { get; set; }
 	}
-#region Extension method for Pixbuf
-	internal static class PixbufExtension
-	{
-		public static unsafe Pixbuf ColorShiftPixbuf (this Pixbuf src, byte shift = 120)
-		{
-			var dest = new Gdk.Pixbuf (src.Colorspace, src.HasAlpha, src.BitsPerSample, src.Width, src.Height);
-
-			byte* src_pixels_orig = (byte*)src.Pixels;
-			byte* dest_pixels_orig = (byte*)dest.Pixels;
-
-			for (int i = 0; i < src.Height; i++) {
-				byte* src_pixels = src_pixels_orig + i * src.Rowstride;
-				byte* dest_pixels = dest_pixels_orig + i * dest.Rowstride;
-
-				for (int j = 0; j < src.Width; j++) {
-					*(dest_pixels++) = PixelClamp (*(src_pixels++) + shift);
-					*(dest_pixels++) = PixelClamp (*(src_pixels++) + shift);
-					*(dest_pixels++) = PixelClamp (*(src_pixels++) + shift);
-
-					if (src.HasAlpha) {
-						*(dest_pixels++) = *(src_pixels++);
-					}
-				}
-			}
-			return dest;
-		}
-
-		static byte PixelClamp (int val)
-		{
-			return (byte)System.Math.Max (0, System.Math.Min (255, val));
-		}
-	}
-#endregion
 }
